% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_dag.R
\name{copula_dag}
\alias{copula_dag}
\title{Copula-based Directed Acyclic Graph}
\usage{
copula_dag(x, alpha = 0.1, pc = TRUE, skeleton = FALSE, fast = TRUE,
  ...)
}
\arguments{
\item{x}{a matrix or data frame with more than two columns
(see \code{\link{pairwise_direction}} for the bivariate case)}

\item{alpha}{a scalar number in (0,1) indicating the level of the test.}

\item{pc}{a logical indicating whether the PC algorithm should be used
to find the skeleton.}

\item{skeleton}{a logical indicating whether the skeleton only should be
returned.}

\item{fast}{a logical indicating whether the memory greedy (but faster)
method should be used.}

\item{...}{additional arguments passed to
\code{\link[rvinecopulib:bicop]{bicop}}}
}
\value{
estimated causal DAG and its adjacency matrix
}
\description{
Copula-based Directed Acyclic Graph
}
\details{
TODO
}
\examples{

set.seed(123)
n <- 500
X <- rnorm(n)
Y <- X ^ 2 + rnorm(n)
Z <- Y ^ 4 + 5 * rnorm(n)
W <- X ^ 3 + 2 * rnorm(n)
dag_data <- cbind(X, Y, Z, W)
trueDAG <- cbind(c(0, 0, 0, 0), c(1, 0, 0, 0), c(0, 1, 0, 0), c(1, 0, 0, 0))

estDAG <- copula_dag(dag_data, alpha = 0.01)

}
