% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulaDAG.R
\docType{package}
\name{copulaDAG}
\alias{copulaDAG}
\alias{copulaDAG-package}
\title{Copula based Directed Acyclic Graphs}
\description{
Copula based Directed Acyclic Graphs
}
\examples{
# TODO
}
\author{
Thibault Vatter
}
\keyword{package}
