% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_direction.R
\name{pairwise_direction}
\alias{pairwise_direction}
\title{Copula-based direction for a single pair}
\usage{
pairwise_direction(x, y = NULL, cop = NULL, ...)
}
\arguments{
\item{x}{a numeric vector, bivariate matrix or data frame.}

\item{y}{\code{NULL} (default) or a vector with compatible dimensions to \code{x}.}

\item{cop}{\code{NULL} (default) or a
\code{\link[rvinecopulib:bicop_dist]{bicop_dist}} object.}

\item{...}{additional arguments passed to
\code{\link[rvinecopulib:bicop]{bicop}}}
}
\value{
TRUE if X -> Y and FALSE if Y -> X
}
\description{
Copula-based direction for a single pair
}
\details{
Assuming that either (a) X := N_X and Y = f(X, N_Y) or (b) Y := N_Y and
X = f(Y, N_X), where N_X and N_Y are independent noise processes,
this copula-based method recovers the direction,
namely (a) X -> Y or (b) Y -> X.
}
\examples{
# Y = X^2 + N_Y with Gaussian noises
n <- 1e3
NX <- rnorm(n)
NY <- rnorm(n)
X <- NX
Y <- X^2 + NY

pairwise_direction(X, Y)
}
