% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_dag.R
\name{copula_dag}
\alias{copula_dag}
\title{Copula-based Directed Acyclic Graph}
\usage{
copula_dag(x, alpha = 0.1, skeleton = FALSE, ...)
}
\arguments{
\item{x}{a matrix or data frame with more than two columns
(see \code{\link{pairwise_direction}} for the bivariate case)}

\item{alpha}{a scalar number in (0,1) indicating the level of the test.}

\item{...}{additional arguments passed to
\code{\link[rvinecopulib:bicop]{bicop}}}

\item{skelecton}{a logical indicating whether the skeleton only should be
returned.}
}
\value{
TODO
}
\description{
Copula-based Directed Acyclic Graph
}
\details{
TODO
}
\examples{
TODO 
}
